Revert "124 clean up"
