+++ src/Pages/Eval/Evals.jsx
@@ -5,9 +5,6 @@ import { useNavigate } from "react-router-dom";
 import { db } from "../../Services/firebase";
 import React, { useEffect, useState } from "react";
 import { Dropdown, InputGroup, Table, Form, Button } from "react-bootstrap";
-import FilterTableHeader from "../../Components/FilterTableHeader";
-import DropdownTableHeaderToggle from "../../Components/DropdownTableHeaderToggle";
-import SortTableHeader from "../../Components/SortTableHeader";
 
 const Evals = () => {
   const [students, setStudents] = useState(null);
@@ -124,6 +121,103 @@ const Evals = () => {
     }
   }
 
+  const DropdownTableHeaderToggle = React.forwardRef(
+    ({ children, onClick }, ref) => (
+      <div
+        className='d-flex'
+        ref={ref}
+        onClick={(e) => {
+          e.preventDefault();
+          onClick(e);
+        }}
+      >
+        {children}
+      </div>
+    ),
+  );
+
+  const FilterTableHeader = React.forwardRef(
+    (
+      {
+        children,
+        style,
+        className,
+        "aria-labelledby": labeledBy,
+        value,
+        valueSetter,
+      },
+      ref,
+    ) => (
+      <div
+        ref={ref}
+        style={style}
+        className={className}
+        aria-labelledby={labeledBy}
+      >
+        <Dropdown.Item>
+          <InputGroup>
+            <Form.Control
+              autoFocus
+              type='text'
+              placeholder='Search'
+              value={value}
+              onChange={(e) => valueSetter(e.target.value)}
+            />
+            <i
+              className='bi bi-x-lg input-group-text'
+              style={{ cursor: "pointer" }}
+              onClick={() => valueSetter("")}
+            />
+          </InputGroup>
+        </Dropdown.Item>
+      </div>
+    ),
+  );
+
+  const ComboTableHeader = React.forwardRef(
+    (
+      {
+        children,
+        style,
+        className,
+        "aria-labelledby": labeledBy,
+        value,
+        valueSetter,
+      },
+      ref,
+    ) => (
+      <div
+        ref={ref}
+        style={style}
+        className={className}
+        aria-labelledby={labeledBy}
+      >
+        {/* <Dropdown.Item>
+          <InputGroup>
+            <Form.Control
+              autoFocus
+              type='text'
+              placeholder='Search'
+              value={value}
+              onChange={(e) => valueSetter(e.target.value)}
+            />
+            <i
+              className='bi bi-x-lg input-group-text'
+              style={{ cursor: "pointer" }}
+              onClick={() => valueSetter("")}
+            />
+          </InputGroup>
+        </Dropdown.Item> */}
+        <Dropdown.Item onClick={() => setTableSort("name_asc")}>
+          A - Z
+        </Dropdown.Item>
+        <Dropdown.Item onClick={() => setTableSort("name_desc")}>
+          Z - A
+        </Dropdown.Item>
+      </div>
+    ),
+  );
+
   const listTable = (
     <Table striped hover>
       <thead>
@@ -133,7 +227,11 @@ const Evals = () => {
               <Dropdown.Toggle as={DropdownTableHeaderToggle}>
                 Student Name {filterIcon("name")}
               </Dropdown.Toggle>
-              <Dropdown.Menu as={SortTableHeader} sortSetter={setTableSort} />
+              <Dropdown.Menu
+                as={ComboTableHeader}
+                value={nameFilter}
+                valueSetter={setNameFilter}
+              />
             </Dropdown>
           </th>
           <th style={{ cursor: "pointer" }}>
