+++ src/Pages/Student/StudentProfilesList.jsx
@@ -220,7 +220,14 @@ const StudentProfilesList = () => {
 
   const FilterTableHeader = React.forwardRef(
     (
-      { style, className, "aria-labelledby": labeledBy, value, valueSetter },
+      {
+        children,
+        style,
+        className,
+        "aria-labelledby": labeledBy,
+        value,
+        valueSetter,
+      },
       ref,
     ) => (
       <div
@@ -250,13 +257,39 @@ const StudentProfilesList = () => {
   );
 
   const ComboTableHeader = React.forwardRef(
-    ({ style, className, "aria-labelledby": labeledBy }, ref) => (
+    (
+      {
+        children,
+        style,
+        className,
+        "aria-labelledby": labeledBy,
+        value,
+        valueSetter,
+      },
+      ref,
+    ) => (
       <div
         ref={ref}
         style={style}
         className={className}
         aria-labelledby={labeledBy}
       >
+        {/* <Dropdown.Item>
+          <InputGroup>
+            <Form.Control
+              autoFocus
+              type='text'
+              placeholder='Search'
+              value={value}
+              onChange={(e) => valueSetter(e.target.value)}
+            />
+            <i
+              className='bi bi-x-lg input-group-text'
+              style={{ cursor: "pointer" }}
+              onClick={() => valueSetter("")}
+            />
+          </InputGroup>
+        </Dropdown.Item> */}
         <Dropdown.Item onClick={() => setTableSort("name_asc")}>
           A - Z
         </Dropdown.Item>
